Change in the severity for 404 HTML code
Checking for Catalan translation is done at domain level. But a domain have some files (one or more). If a file of a domain is not translated, this file doesn't exist and we obtain a 404 HTML code. With this PR, I suggest a change in the severity for 404 HTML code, from error to info severity. This will help to maintain clean the error log, while we will see not found files in the info log.